+++ src/components/Alert/Alert.types.ts
@@ -1,0 +1,5 @@
+import { BrandColor, FontColor, IconName } from '../../types';
+
+export type AlertVariant = 'info' | 'success' | 'warning' | 'danger' | 'default';
+
+export type AlertAttributes = { icon: IconName; color: FontColor; background: BrandColor; };