+++ src/components/Modal/Modal.tsx
@@ -3,7 +3,7 @@ import React, {
   RefObject,
   forwardRef,
 } from 'react';
-import { DialogOverlay, DialogContent } from '@reach/dialog';
+import { DialogOverlay, DialogContent } from '@palmetto/dialog';
 import classNames from 'classnames';
 import { CssOverflowValue } from '../../types';
 import { getDimensionCss } from '../../lib/getDimensionCss';
@@ -28,13 +28,19 @@ export interface ModalProps {
    */
   ariaLabelledBy?: string;
   /**
-   * Contents of the button.
+   * Contents of the dialog.
    */
   children?: ReactNode;
   /**
-   * Additional ClassNames to add to button.
+   * Additional ClassNames to add to dialog.
    */
   className?: string;
+  /**
+   * The ref of the container where the dialog will be inserted into the DOM.
+   * By default, Modals are inserted in the document.body, but if need be they can
+   * be scoped as necessary.
+   */
+  containerRef?: React.RefObject<Node>;
   /**
    * At mobile viewport widths, the modal will take up the fullscreen
    */
@@ -56,7 +62,7 @@ export interface ModalProps {
    */
   maxWidth?: BoxProps['maxWidth'];
   /**
-   * Function that is called whenever the user hits "Esacape" key or clicks outside the modal.
+   * Function that is called whenever the user hits "Escape" key or clicks outside the modal.
    */
   onDismiss: (event?: React.SyntheticEvent) => void;
   /**
@@ -76,6 +82,7 @@ const ModalBaseComponent: React.FC<ModalProps> = forwardRef<HTMLDivElement, Moda
     allowPinchZoom = false,
     children,
     className,
+    containerRef = undefined,
     fullScreenMobile = false,
     initialFocusRef,
     isOpen,
@@ -88,9 +95,13 @@ const ModalBaseComponent: React.FC<ModalProps> = forwardRef<HTMLDivElement, Moda
 ) => {
   const maxWidthCss = getDimensionCss('mw', maxWidth);
 
-  const overylayClassnames = classNames(styles.overlay, {
-    fullscreen: fullScreenMobile,
-  });
+  const overlayClassnames = classNames(
+    styles.overlay,
+    styles.modal,
+    {
+      fullscreen: fullScreenMobile,
+    },
+  );
   const contentClassnames = classNames(
     styles['modal-content'],
     className,
@@ -102,7 +113,8 @@ const ModalBaseComponent: React.FC<ModalProps> = forwardRef<HTMLDivElement, Moda
 
   return (
     <DialogOverlay
-      className={overylayClassnames}
+      className={overlayClassnames}
+      containerRef={containerRef}
       allowPinchZoom={allowPinchZoom}
       isOpen={isOpen}
       onDismiss={onDismiss}