+++ src/components/SelectInput/SelectInput.tsx
@@ -156,6 +156,7 @@ export const SelectInput: FC<SelectInputProps> = ({
   const responsiveClasses = generateResponsiveClasses('size', size);
 
   const wrapperClasses = classNames(
+    'palmetto-components__variables__form-control',
     'select-input-wrapper',
     className,
     ...responsiveClasses.map(c => (styles[c])),