+++ src/components/TextareaInput/TextareaInput.VisualTests.stories.tsx
@@ -23,6 +23,7 @@ ResponsiveSize.args = {
     desktop: 'lg',
     hd: 'sm',
   },
-  value: 'responsive',
+  value: 'Responsive size',
+  label: 'Textarea Input Visual Test',
 };
 ResponsiveSize.parameters = RESPONSIVE_STORY;